--- CHANGELOG.md
@@ -1,29 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-## [0.1.6] - 2024-10-10
-
-### Added
-- Command line interface with `wpt` command
-- Automated PyPI deployment with OpenID Connect
-- Sphinx documentation with CLI reference
-- Regression tests with snapshot validation
-- PyPI deployment documentation
-
-### Changed
-- Refactored `nw_tools` to use static methods
-- Improved GitHub Actions workflows
-- Enhanced test coverage reporting
-
-### Fixed
-- Windows compatibility for LCR fitting
-- Sphinx documentation configuration
-- Test workflow parameters
-
-## [0.1.5] - Previous Release
-
-### Added
-- Initial release with core functionality
-- Basic wireless power analysis tools
-- S-parameter processing capabilities
