+++ docs/source/solver.md
