+++ tests/test_solver.py
